# Introduce `TreeNodeMutator` for rewriting `TreeNode`s in place, change optimizer to rewrite `LogicalPlan` in place - 10% faster planning time