… datafusion-examples/examples/rewrite_expr.rs
… datafusion/core/tests/optimizer_integration.rs
… datafusion/optimizer/src/decorrelate_predicate_subquery.rs
… datafusion/optimizer/src/eliminate_filter.rs
… datafusion/optimizer/src/eliminate_limit.rs
… datafusion/optimizer/src/eliminate_nested_union.rs
… datafusion/optimizer/src/eliminate_one_union.rs
… datafusion/optimizer/src/eliminate_outer_join.rs
… datafusion/optimizer/src/extract_equijoin_predicate.rs
… datafusion/optimizer/src/filter_null_join_keys.rs
… datafusion/optimizer/src/optimize_projections.rs
… datafusion/optimizer/src/optimizer.rs
… datafusion/optimizer/src/propagate_empty_relation.rs
… datafusion/optimizer/src/push_down_filter.rs
… datafusion/optimizer/src/push_down_limit.rs
… datafusion/optimizer/src/push_down_projection.rs
… datafusion/optimizer/src/replace_distinct_aggregate.rs
… datafusion/optimizer/src/scalar_subquery_to_join.rs
… datafusion/optimizer/src/single_distinct_to_groupby.rs
… datafusion/optimizer/src/test/mod.rs
… datafusion/optimizer/tests/optimizer_integration.rs
… datafusion/sqllogictest/test_files/predicates.slt
… datafusion/sqllogictest/test_files/subquery.slt
… datafusion/sqllogictest/test_files/timestamps.slt