Note this PR looks large, but it is mostly comments and changes of `&plan` to `plan` in tests

## Which issue does this PR close?

Part of https://github.com/apache/arrow-datafusion/issues/9637 (based on some ideas from https://github.com/apache/arrow-datafusion/pull/9708). 🙏 @jayzhan211 

closes https://github.com/apache/arrow-datafusion/pull/9768

## Rationale for this change
See https://github.com/apache/arrow-datafusion/issues/9637

TLDR is that the optimizer currently copies `LogicalPlan`s many times unnecessarily (I think it is quadratic if not worse in the number of tree nodes). This is both slow and requires many memory allocations. 

Note we can do even better, but I want to keep the PR sizes as small as possible. 

## What changes are included in this PR?


1. Add in place `mutate` / `Mutator` APIs to `TreeNode`
2. Implement `mutate` for `LogicalPlan` (aka `&mut LogicalPlan`) 
1. Rewrite `Optimizer` to use `TreeNode` API

I am trying to keep this PR relatively small, and I have other improvements in mind for follow on PRs (see below). This PR is now finally feasible thanks to the really nice TreeNode cleanup from @peter-toth  and @berkaysynnada in https://github.com/apache/arrow-datafusion/pull/8891 


## Are these changes tested?
Functionally covered by existing tests, and I measured performance using the benchmarks -- I see an mprovement of about 10% across TPCH queries.

<details><summary>Details</summary>
<p>

```
++ critcmp main optimizer_tree_node
++ critcmp main optimizer_tree_node
group                                         main                                    optimizer_tree_node
-----                                         ----                                    -------------------
logical_aggregate_with_join                   1.01  1287.8±105.61µs        ? ?/sec    1.00   1274.7±7.85µs        ? ?/sec
logical_plan_tpch_all                         1.00     17.3±0.16ms        ? ?/sec     1.00     17.3±0.14ms        ? ?/sec
logical_select_all_from_1000                  1.00     93.4±1.40ms        ? ?/sec     1.02     95.3±0.34ms        ? ?/sec
logical_select_one_from_700                   1.00    740.6±8.75µs        ? ?/sec     1.00   743.8±17.11µs        ? ?/sec
logical_trivial_join_high_numbered_columns    1.00   789.6±22.14µs        ? ?/sec     1.00   792.5±19.06µs        ? ?/sec
logical_trivial_join_low_numbered_columns     1.00   755.3±13.71µs        ? ?/sec     1.00   759.0±18.47µs        ? ?/sec
physical_plan_tpch_all                        1.11    133.8±0.87ms        ? ?/sec     1.00    120.7±0.70ms        ? ?/sec
physical_plan_tpch_q1                         1.15      7.7±0.15ms        ? ?/sec     1.00      6.7±0.05ms        ? ?/sec
physical_plan_tpch_q10                        1.10      6.3±0.05ms        ? ?/sec     1.00      5.7±0.04ms        ? ?/sec
physical_plan_tpch_q11                        1.09      5.0±0.03ms        ? ?/sec     1.00      4.6±0.03ms        ? ?/sec
physical_plan_tpch_q12                        1.08      4.1±0.03ms        ? ?/sec     1.00      3.8±0.03ms        ? ?/sec
physical_plan_tpch_q13                        1.09      2.7±0.03ms        ? ?/sec     1.00      2.5±0.02ms        ? ?/sec
physical_plan_tpch_q14                        1.05      3.4±0.02ms        ? ?/sec     1.00      3.3±0.02ms        ? ?/sec
physical_plan_tpch_q16                        1.12      5.3±0.04ms        ? ?/sec     1.00      4.7±0.04ms        ? ?/sec
physical_plan_tpch_q17                        1.11      4.9±0.04ms        ? ?/sec     1.00      4.4±0.03ms        ? ?/sec
physical_plan_tpch_q18                        1.11      5.4±0.04ms        ? ?/sec     1.00      4.9±0.03ms        ? ?/sec
physical_plan_tpch_q19                        1.06     10.1±0.09ms        ? ?/sec     1.00      9.5±0.07ms        ? ?/sec
physical_plan_tpch_q2                         1.14     12.2±0.09ms        ? ?/sec     1.00     10.7±0.10ms        ? ?/sec
physical_plan_tpch_q20                        1.14      6.4±0.05ms        ? ?/sec     1.00      5.6±0.04ms        ? ?/sec
physical_plan_tpch_q21                        1.11      9.5±0.08ms        ? ?/sec     1.00      8.6±0.18ms        ? ?/sec
physical_plan_tpch_q22                        1.10      4.7±0.04ms        ? ?/sec     1.00      4.2±0.03ms        ? ?/sec
physical_plan_tpch_q3                         1.08      4.2±0.04ms        ? ?/sec     1.00      3.9±0.03ms        ? ?/sec
physical_plan_tpch_q4                         1.09      3.3±0.02ms        ? ?/sec     1.00      3.0±0.03ms        ? ?/sec
physical_plan_tpch_q5                         1.08      6.1±0.04ms        ? ?/sec     1.00      5.6±0.04ms        ? ?/sec
physical_plan_tpch_q6                         1.05      2.0±0.02ms        ? ?/sec     1.00  1944.8±12.65µs        ? ?/sec
physical_plan_tpch_q7                         1.11      8.6±0.09ms        ? ?/sec     1.00      7.8±0.06ms        ? ?/sec
physical_plan_tpch_q8                         1.12     12.1±0.10ms        ? ?/sec     1.00     10.8±0.09ms        ? ?/sec
physical_plan_tpch_q9                         1.13      9.2±0.07ms        ? ?/sec     1.00      8.1±0.11ms        ? ?/sec
physical_select_all_from_1000                 1.13    684.2±1.45ms        ? ?/sec     1.00    606.5±1.42ms        ? ?/sec
physical_select_one_from_700                  1.00      4.2±0.03ms        ? ?/sec     1.01      4.2±0.03ms        ? ?/sec
```

</p>
</details> 

## Are there any user-facing changes?
1. Faster planning
3. Ability to rewrite  `LogicalPlan` in place using `TreeNode` API


## Planned Follow Ons:
Here are follow on tickets that will improve things even more
- [ ] Add API to `OptimizerRule` to rewrite in place (and avoid yet another copy)
- [ ] Add an example of how to use the "in place tree write" API
- [ ] Supporting rewriting `Expr`s in place in addition to `LogicalPlan` (implement `TreeNodeMutator` -- specifically ExprRewriter should be a big win)
- [ ] API to support in place rewrites of `UserDefinedLogicalPlan`
- [ ] Change other uses of `TreeNodeRewriter` to `TreeNodeMutator` (e.g. `ExprRewriter`, etc)