@alamb 2024-03-29
Update here is that I think this PR now has the basic code to do in-place updates of TreeNodes in general and `LogicalPlan`s in particular 🎉 



TODOs:
- [ ] Run performance benchmarks to show this helps 
- [ ] Complete edge case implementations (DDL, user defined nodes), polish PR description

Follow ons:
- [ ] File ticket for supporting rewriting `Expr`s in place (implement `TreeNodeMutator` -- specifically ExprRewriter should be a big win)
- [ ] File ticket for: Better OptimizerRule API to support rewriting in place
- [ ] File ticket for: API to support rewrites of UserDefinedLogicalPlan
- [ ] File ticket for: change additional uses of `TreeNodeRewriter` to `TreeNodeMutator`
- [ ] File ticket for: change additional uses of TreeNodeRewriter to TreeNodeMutator
- [ ] File ticket for  an example of how to use the "in place tree write" API