@alamb 2024-03-30
Updated description with performance results (about a 10% improvement for this PR alone). I think we can squeeze a bunch more out of this approach with the additional changes to rewrite in place (will file follow on tickets). 

Proceeding to get this PR ready for review