@peter-toth 2024-04-02
> So how about this for a plan:
> 
> 1. I'll make a PR that changes LogicalPlan::`map_children` to avoid copies if possible (which should be beneficial in many places in DataFusion)
> 2. I'll create a PR that changes the Optimizer to use a TreeNodeRewriter (rather than TreeNodeMutator) which will avoid even more copies after 1 is complete.
> 
> We can then compare the performance of the TreeNodeRewriter to this PR with TreeNodeMutator and decide and see if the additional TreeNodeMutator API is worth the extra API surface / migration.

Sounds very good to me, I'm really curious how much improvement the changes can bring.