@alamb 2024-04-04
Per https://github.com/apache/arrow-datafusion/pull/9780#discussion_r1552005102 I think we can accomplish the same thing without a mutator at this time, so closing this PR